+++ speeduino/utils.h
@@ -1,12 +1,12 @@
 /*
 These are some utility functions and variables used through the main code
-*/ 
+*/
 #ifndef UTILS_H
 #define UTILS_H
 
 #include <Arduino.h>
 
-int freeRam ();
+uint16_t freeRam ();
 void setPinMapping(byte boardID);
 unsigned int PW();
 unsigned int PW_SD();
